# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: JobPing API
  description: |
    JobPing is a sophisticated job matching platform for early-career professionals.
    This API provides comprehensive job matching, user management, and analytics capabilities.
    
    ## Key Features
    - AI-powered job matching with GPT-4
    - Rule-based fallback matching
    - Real-time job scraping from multiple sources
    - User preference management
    - Email automation and notifications
    - Payment processing with Stripe
    - Comprehensive analytics and monitoring
    
    ## Authentication
    Most endpoints require authentication via JWT tokens or API keys.
    Public endpoints are clearly marked.
    
    ## Rate Limiting
    API requests are rate-limited per endpoint to ensure fair usage:
    - Public endpoints: 20 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Resource-intensive endpoints: 2-5 requests/minute
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
    See the Error Response schema for details.
  version: 1.0.0
  contact:
    name: JobPing Support
    email: support@jobping.ai
    url: https://jobping.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://jobping.ai/api
    description: Production server
  - url: https://staging.jobping.ai/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /send-scheduled-emails:
    post:
      tags:
        - Email Automation
      summary: Send Scheduled Email Campaigns
      description: |
        Sends scheduled email campaigns to users with their job matches.
        This endpoint is typically called by automated systems.
        
        **Rate Limit**: 1 request per minute (system endpoint)
      operationId: sendScheduledEmails
      security:
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                campaignType:
                  type: string
                  enum: [daily, weekly, immediate]
                  default: daily
                maxUsers:
                  type: integer
                  minimum: 1
                  maximum: 1000
                  default: 100
              example:
                campaignType: daily
                maxUsers: 100
      responses:
        '200':
          description: Emails sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  emailsSent:
                    type: integer
                    example: 87
                  usersWithMatches:
                    type: integer
                    example: 76
                  usersWithoutMatches:
                    type: integer
                    example: 11
                  processingTime:
                    type: number
                    example: 12.5
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /create-checkout-session:
    post:
      tags:
        - Payments
      summary: Create Stripe Checkout Session
      description: |
        Creates a Stripe checkout session for subscription payments.
        Supports both monthly and annual billing cycles.
      operationId: createCheckoutSession
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - priceId
                - successUrl
                - cancelUrl
              properties:
                priceId:
                  type: string
                  description: Stripe price ID for the subscription
                  example: "price_1234567890"
                successUrl:
                  type: string
                  format: uri
                  description: URL to redirect after successful payment
                  example: "https://jobping.ai/billing/success"
                cancelUrl:
                  type: string
                  format: uri
                  description: URL to redirect after cancelled payment
                  example: "https://jobping.ai/billing/cancel"
                customerEmail:
                  type: string
                  format: email
                  description: Customer email for the checkout session
                  example: "user@example.com"
      responses:
        '200':
          description: Checkout session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    example: "cs_1234567890"
                  url:
                    type: string
                    format: uri
                    example: "https://checkout.stripe.com/pay/cs_1234567890"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for system authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid email format"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        requestId:
          type: string
          example: "req_1234567890"
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Bad Request"
            message: "Invalid request parameters"
            code: "BAD_REQUEST"
            timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Authentication required"
            code: "UNAUTHORIZED"
            timestamp: "2024-01-15T10:30:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-15T10:30:00Z"